virtio_socket: protect the active_sockets counter
Prevent a race condition when the active_sockets is checked when the
virtio_socket module is unloaded.

A socket could be created between the time the value is acquired and
checked to see if it's greater than zero allowing the module to be
unloaded with active sockets.

I'm not super happy with the solution...