vsock: use epoch to manage PCB memory
Use epoch to manage PCB lifetimes. Remove the transport_lock from the
send code path. The same was being used to prevent the device to use PCB
after they being released. Now a per PCB lock is acquired in the send
code path.

Change the PCB lock from a mtx to an sx lock to workaround a LOR. The
transport_lock is still acquired when a socket is detached and when the
virtio_socket device is unloaded so we can check if there are active
sockets and prevent the driver being unloaded.